+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.1.98](https://github.com/edwardcalderon/ComputeIntelligenceGraph/compare/v0.1.97...v0.1.98) (2026-03-24)
+
+### Detected changes
+- packages: 2 changed file(s)
+  - packages/api/src/index.ts
+  - packages/api/src/security-testing.test.ts
+
+
+## [0.1.98](https://github.com/edwardcalderon/ComputeIntelligenceGraph/compare/v0.1.97...v0.1.98) (2026-03-24)
+
+
+
+
+
 ## [0.1.97](https://github.com/edwardcalderon/ComputeIntelligenceGraph/compare/v0.1.96...v0.1.97) (2026-03-24)
 
 ### Detected changes